# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CaseBinder AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaseCaseType = (typeof CaseCaseType)[keyof typeof CaseCaseType];

export const CaseCaseType = {
  custody: "custody",
  divorce: "divorce",
  contract_dispute: "contract_dispute",
  employment: "employment",
  landlord_tenant: "landlord_tenant",
  other: "other",
} as const;

export interface Case {
  id: number;
  userId: string;
  name: string;
  caseType: CaseCaseType;
  parties: string;
  attorneyName?: string | null;
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateCaseBodyCaseType =
  (typeof CreateCaseBodyCaseType)[keyof typeof CreateCaseBodyCaseType];

export const CreateCaseBodyCaseType = {
  custody: "custody",
  divorce: "divorce",
  contract_dispute: "contract_dispute",
  employment: "employment",
  landlord_tenant: "landlord_tenant",
  other: "other",
} as const;

export interface CreateCaseBody {
  name: string;
  caseType: CreateCaseBodyCaseType;
  parties: string;
  attorneyName?: string | null;
  description?: string | null;
}

export type UpdateCaseBodyCaseType =
  (typeof UpdateCaseBodyCaseType)[keyof typeof UpdateCaseBodyCaseType];

export const UpdateCaseBodyCaseType = {
  custody: "custody",
  divorce: "divorce",
  contract_dispute: "contract_dispute",
  employment: "employment",
  landlord_tenant: "landlord_tenant",
  other: "other",
} as const;

export interface UpdateCaseBody {
  name?: string;
  caseType?: UpdateCaseBodyCaseType;
  parties?: string;
  attorneyName?: string | null;
  description?: string | null;
}

export interface CaseStats {
  caseId: number;
  evidenceCount: number;
  timelineEventCount: number;
  suggestedEventCount: number;
  exportCount: number;
  lastExportStatus?: string | null;
}

export type EvidenceProcessingStatus =
  (typeof EvidenceProcessingStatus)[keyof typeof EvidenceProcessingStatus];

export const EvidenceProcessingStatus = {
  pending: "pending",
  processing: "processing",
  processed: "processed",
  failed: "failed",
} as const;

export interface Evidence {
  id: number;
  caseId: number;
  filename: string;
  fileType: string;
  objectPath: string;
  fileSize?: number | null;
  detectedDate?: string | null;
  processingStatus: EvidenceProcessingStatus;
  textPreview?: string | null;
  tags: string[];
  people: string[];
  uploadedAt: string;
}

export interface CreateEvidenceBody {
  filename: string;
  fileType: string;
  objectPath: string;
  fileSize?: number | null;
  detectedDate?: string | null;
  tags?: string[];
  people?: string[];
}

export type UpdateEvidenceBodyProcessingStatus =
  (typeof UpdateEvidenceBodyProcessingStatus)[keyof typeof UpdateEvidenceBodyProcessingStatus];

export const UpdateEvidenceBodyProcessingStatus = {
  pending: "pending",
  processing: "processing",
  processed: "processed",
  failed: "failed",
} as const;

export interface UpdateEvidenceBody {
  detectedDate?: string | null;
  processingStatus?: UpdateEvidenceBodyProcessingStatus;
  textPreview?: string | null;
  tags?: string[];
  people?: string[];
}

export interface TimelineEvent {
  id: number;
  caseId: number;
  eventDate: string;
  eventTime?: string | null;
  title: string;
  description?: string | null;
  attorneyNote?: string | null;
  people: string[];
  tags: string[];
  createdAt: string;
}

export interface CreateTimelineEventBody {
  eventDate: string;
  eventTime?: string | null;
  title: string;
  description?: string | null;
  attorneyNote?: string | null;
  people?: string[];
  tags?: string[];
}

export interface UpdateTimelineEventBody {
  eventDate?: string;
  eventTime?: string | null;
  title?: string;
  description?: string | null;
  attorneyNote?: string | null;
  people?: string[];
  tags?: string[];
}

export interface AttachEvidenceBody {
  evidenceId: number;
}

export type TranscriptSource =
  (typeof TranscriptSource)[keyof typeof TranscriptSource];

export const TranscriptSource = {
  typed: "typed",
  microphone: "microphone",
} as const;

export interface Transcript {
  id: number;
  caseId: number;
  content: string;
  source: TranscriptSource;
  createdAt: string;
  updatedAt: string;
}

export type CreateTranscriptBodySource =
  (typeof CreateTranscriptBodySource)[keyof typeof CreateTranscriptBodySource];

export const CreateTranscriptBodySource = {
  typed: "typed",
  microphone: "microphone",
} as const;

export interface CreateTranscriptBody {
  content: string;
  source: CreateTranscriptBodySource;
}

export interface UpdateTranscriptBody {
  content?: string;
}

export interface GenerateSummaryBody {
  transcriptId: number;
  additionalContext?: string | null;
}

export interface GeneratedSummary {
  summary: string;
  suggestedTitle?: string | null;
}

export interface SuggestEventsBody {
  transcriptId: number;
}

export type SuggestedEventConfidenceLevel =
  (typeof SuggestedEventConfidenceLevel)[keyof typeof SuggestedEventConfidenceLevel];

export const SuggestedEventConfidenceLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type SuggestedEventStatus =
  (typeof SuggestedEventStatus)[keyof typeof SuggestedEventStatus];

export const SuggestedEventStatus = {
  pending: "pending",
  accepted: "accepted",
  ignored: "ignored",
} as const;

export interface SuggestedEvent {
  id: number;
  caseId: number;
  estimatedDate?: string | null;
  title: string;
  description: string;
  people: string[];
  confidenceLevel: SuggestedEventConfidenceLevel;
  suggestedSearchTerms: string[];
  status: SuggestedEventStatus;
  createdAt: string;
}

export interface CaseSummary {
  id: number;
  caseId: number;
  content: string;
  updatedAt: string;
}

export interface SaveCaseSummaryBody {
  content: string;
}

export type ExportExportType =
  (typeof ExportExportType)[keyof typeof ExportExportType];

export const ExportExportType = {
  pdf: "pdf",
  zip: "zip",
} as const;

export type ExportStatus = (typeof ExportStatus)[keyof typeof ExportStatus];

export const ExportStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Export {
  id: number;
  caseId: number;
  exportType: ExportExportType;
  status: ExportStatus;
  objectPath?: string | null;
  createdAt: string;
}

export type CreateExportBodyExportType =
  (typeof CreateExportBodyExportType)[keyof typeof CreateExportBodyExportType];

export const CreateExportBodyExportType = {
  pdf: "pdf",
  zip: "zip",
} as const;

export interface CreateExportBody {
  exportType: CreateExportBodyExportType;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListEvidenceParams = {
  search?: string;
  fileType?: string;
  dateFrom?: string;
  dateTo?: string;
  tag?: string;
};

export type AttachEvidenceToEvent201 = {
  success: boolean;
};

export type IgnoreSuggestedEvent200 = {
  success: boolean;
};