# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CaseBinder AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases for the authenticated user
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a specific case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseBody"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cases/{caseId}/stats:
    get:
      operationId: getCaseStats
      tags: [cases]
      summary: Get case stats (evidence count, timeline events, export status)
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStats"

  # ── Evidence ─────────────────────────────────────────────────────────────────
  /cases/{caseId}/evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: fileType
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of evidence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidence"
    post:
      operationId: createEvidence
      tags: [evidence]
      summary: Register evidence metadata after upload
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvidenceBody"
      responses:
        "201":
          description: Created evidence record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
  /cases/{caseId}/evidence/{evidenceId}:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: Get a specific evidence item
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
    patch:
      operationId: updateEvidence
      tags: [evidence]
      summary: Update evidence metadata
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvidenceBody"
      responses:
        "200":
          description: Updated evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
    delete:
      operationId: deleteEvidence
      tags: [evidence]
      summary: Delete evidence
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Timeline Events ──────────────────────────────────────────────────────────
...
  /cases/{caseId}/exports:
    get:
      operationId: listExports
      tags: [exports]
      summary: List exports for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of exports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Export"
    post:
      operationId: createExport
      tags: [exports]
      summary: Create/trigger an export
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExportBody"
      responses:
        "201":
          description: Export created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"

  # ── Storage ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Case:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        caseType:
          type: string
          enum: [custody, divorce, contract_dispute, employment, landlord_tenant, other]
        parties:
          type: string
        attorneyName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, caseType, parties, createdAt, updatedAt]
    CreateCaseBody:
      type: object
      properties:
        name:
          type: string
        caseType:
          type: string
          enum: [custody, divorce, contract_dispute, employment, landlord_tenant, other]
        parties:
          type: string
        attorneyName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required: [name, caseType, parties]
    UpdateCaseBody:
      type: object
      properties:
        name:
          type: string
        caseType:
          type: string
          enum: [custody, divorce, contract_dispute, employment, landlord_tenant, other]
        parties:
          type: string
        attorneyName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    CaseStats:
      type: object
      properties:
        caseId:
          type: integer
        evidenceCount:
          type: integer
        timelineEventCount:
          type: integer
        suggestedEventCount:
          type: integer
        exportCount:
          type: integer
        lastExportStatus:
          type: string
          nullable: true
      required: [caseId, evidenceCount, timelineEventCount, suggestedEventCount, exportCount]
    Evidence:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        objectPath:
          type: string
        fileSize:
          type: integer
          nullable: true
        detectedDate:
          type: string
          nullable: true
        processingStatus:
          type: string
          enum: [pending, processing, processed, failed]
        textPreview:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        people:
          type: array
          items:
            type: string
        uploadedAt:
          type: string
      required: [id, caseId, filename, fileType, objectPath, processingStatus, tags, people, uploadedAt]
    CreateEvidenceBody:
      type: object
      properties:
        filename:
          type: string
        fileType:
          type: string
        objectPath:
          type: string
        fileSize:
          type: integer
          nullable: true
        detectedDate:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        people:
          type: array
          items:
            type: string
      required: [filename, fileType, objectPath]
    UpdateEvidenceBody:
      type: object
      properties:
        detectedDate:
          type: string
          nullable: true
        processingStatus:
          type: string
          enum: [pending, processing, processed, failed]
        textPreview:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        people:
          type: array
          items:
            type: string
...
    Export:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        exportType:
          type: string
          enum: [pdf, zip]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        objectPath:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, caseId, exportType, status, createdAt]
    CreateExportBody:
      type: object
      properties:
        exportType:
          type: string
          enum: [pdf, zip]
      required: [exportType]
...